refactor into hook
